[Feat] [SCRUM-204] Game socket gateway 구현 #160
